Fixed damage taken lesser than defense bug, Added health bar to all the small monsters, added damage indicators to all the small monsters and also the character, made the health bar and mana bar of character work, added 2 new skills (revitalize[heal 30% of max hp, cd of 15s] & rejuvinate[regen 30% of max mana, cd 15s]), Added in a new type of monster called hell hand, fixed UI scaling of the stages

+++ GPN_Assignment/Assets/Scripts/Monsters/Boss.cs
@@ -28,7 +28,7 @@ public class Boss : MonoBehaviour
     public Transform player;
     public LayerMask playerLayer;
     public Collider2D playerCollider;
-    int playerHealth;
+    float playerHealth;
     float distToPlayer;
     public int range;
